Introduce GlobalVariable as scoped wrapper for global variables
