+++ src/Phan/Language/Element/PassByReferenceVariable.php
@@ -20,17 +20,12 @@ use Phan\Language\UnionType;
 class PassByReferenceVariable extends Variable
 {
 
+    use ElementProxyTrait;
+
     /**
      * @var Variable the parameter which accepts references
      */
     private $parameter;
-
-    /**
-     * The element that was passed in as an argument (e.g. variable or static property)
-     * @var TypedElement|UnaddressableTypedElement
-     * TODO: Make a common interface which has methods implemented
-     */
-    private $element;
 
     /**
      * @var ?CodeBase set to a CodeBase if $element is a Property, for type checking
@@ -101,41 +96,6 @@ class PassByReferenceVariable extends Variable
         $this->element->setUnionType($type->eraseRealTypeSetRecursively());
     }
 
-    public function getFlags(): int
-    {
-        return $this->element->getFlags();
-    }
-
-    public function getFlagsHasState(int $bits): bool
-    {
-        return $this->element->getFlagsHasState($bits);
-    }
-
-    public function setFlags(int $flags): void
-    {
-        $this->element->setFlags($flags);
-    }
-
-    public function getPhanFlags(): int
-    {
-        return $this->element->getPhanFlags();
-    }
-
-    public function getPhanFlagsHasState(int $bits): bool
-    {
-        return $this->element->getPhanFlagsHasState($bits);
-    }
-
-    public function setPhanFlags(int $phan_flags): void
-    {
-        $this->element->setPhanFlags($phan_flags);
-    }
-
-    public function getFileRef(): FileRef
-    {
-        return $this->element->getFileRef();
-    }
-
     /**
      * Returns the context where this reference was created.
      * This is currently only available for references to properties.
@@ -153,14 +113,5 @@ class PassByReferenceVariable extends Variable
     public function isPHPInternal(): bool
     {
         return $this->element instanceof AddressableElement && $this->element->isPHPInternal();
-    }
-
-    /**
-     * Get the argument passed in to this object.
-     * @return TypedElement|UnaddressableTypedElement
-     */
-    public function getElement()
-    {
-        return $this->element;
     }
 }
